app/service/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tissuelab-ai-service"
version = "0.1.8"
description = "AI Service for TissueLab"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "async-timeout==4.0.3",
    "attrs==24.2.0",
    "certifi==2024.8.30",
    "charset-normalizer==3.4.0",
    "fastapi==0.109.1",
    # "firebase-admin",  # Removed for open source
    "frozenlist==1.5.0",
    # "google-cloud-firestore",  # Removed for open source
    # "google-cloud-storage",  # Removed for open source
    "h5py==3.12.1",
    "idna==3.10",
    "importlib_metadata==8.5.0",
    "importlib_resources==6.4.5",
    "matplotlib==3.9.2",
    "multidict==6.1.0",
    "nibabel",
    "numba",
    "numpy==1.26.4",
    "numcodecs==0.12.1",
    "ome-types==0.5.2",
    "ome-zarr==0.9.0",
    "opencv-python",
    "pillow==11.0.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.0.3",
    "PyJWT>=2.8.0",
    "python-multipart",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "requests==2.32.3",
    "scipy==1.13.1",
    "six==1.16.0",
    "tifffile==2024.8.30",
    "tiffslide==2.4.0",
    "typing_extensions==4.12.2",
    "urllib3==1.26.20",
    "uvicorn==0.27.0",
    "websockets>=11.0.3",
    "yarl==1.17.1",
    "zarr==2.17.2",
    "pydicom",
    "pylibCZIrw; sys_platform == 'win32'",
    "pyisyntax",
    "pyvips[binary]",
    "openai",
    "czifile",
    "tissuelab_sdk==0.1.8",
    "celery",
    "pyzmq",
]
